starting on converting profile to be fully tailwind
