+++ frontend/src/pages/Popup.tsx
@@ -1,24 +1,43 @@
 import '../styles/global.css'
-import React from 'react'
-import {useNavigate} from 'react-router-dom'
+import React from "react"
 
-export function Popup(){
-
-    const navigate = useNavigate();
-
-    return (
-        <div className="min-h-screen w-full bg-sky-50 flex items-center justify-center p-6">
-
-        <div className="relative w-full max-w-sm">
-
-        <img src="/robot.png" alt="robot" className="absolute -top-28 left-1/2 -translate-x-1/2 w-44 z-10"></img>
-
-    
-
-        </div>
-
-
-        </div>
-    );
+interface PopupProps{
+    onMath? : () => void;
+    onProg? : () => void;
+    onCancel? : () => void;
 }
 
+
+const Popup: React.FC<PopupProps> = ({onMath, onProg, onCancel}) => {
+
+    return(
+        <div className="min-h-screen w-full bg-[var(--secondary)] flex items-center justify-center p-6 bg-[url(./robot.png)] bg-center bg-no-repeat bg-size-[auto_800px]">
+            <div className="relative w-xl">
+                {/* <img src='./robot.png' alt="robot" className="relative w-full z-0 inset-0"></img> */}
+
+                <div className="bg-[var(--secondary)] w-full border rounded-3xl pt-12 pb-8 text-center shadow-lg absolute -top-9 h-124">
+                    <h1 className="text-6xl font-[var(--heading)] text-[var(--secondary-text)] font-extrabold ">
+                        Choose a Topic
+                    </h1>
+                    <h2 className="text-2xl font-[var(--heading)] font-medium text-[var(--secondary-text)] text-center justify-center mt-5">What would you like to be challenged on?</h2>
+                
+                <div className="grid grid-flow-col grid-cols-2 gap-4 ml-8 mr-8 mt-16 h-45">
+                    <div className="bg-[var(--secondary)] rounded-3xl text-center text-[var(--secondary-text)] shadow-lg mr-2">
+                        <h1 className="text-3.3xl text-[var(--secondary-text)] font-[var(--heading)] font-bold mt-5">+ -</h1>
+                        <h2 className="text-3.3xl text-[var(--secondary-text)] font-[var(--heading)] font-bold mt-3" onClick={onMath}>Math</h2>
+                        </div>
+                    <div className="bg-[var(--secondary)] rounded-3xl text-center text-[var(--secondary-text)] shadow-lg ml-2">
+                        <h1 className="text-3.3xl text-[var(--secondary-text)] font-[var(--heading)] font-bold mt-5">{"</>"}</h1>
+                        <h2 className="text-3.3xl text-[var(--secondary-text)] font-[var(--heading)] font-bold mt-3" onClick={onProg}>Programming</h2>
+                    </div>
+                </div>
+                <div className="text-3.3xl text-black font-[var(--heading)] font-extrabold underline mt-5" onClick={onCancel}>Cancel</div>
+                </div>
+            </div>
+        </div>
+    );
+
+};
+
+
+export default Popup;
